game/console: Rearrange status windows, add message windows
Instead of having 3 different windows which show only few pieces of
information, there is now a single window which shows more information
to the user. Also, it's now possible to display a message on the screen
as well as to prompt the user to make a decision (by showing a simple
menu and asking polar questions). Additionally, some deprecated code has
been removed.